address feedback and reorganize files in testdata
