+++ pkg/webhook/operatingsystemconfig/ensurer_test.go
@@ -6,7 +6,6 @@ package operatingsystemconfig_test
 
 import (
 	"context"
-	_ "embed"
 	"fmt"
 
 	gcontext "github.com/gardener/gardener/extensions/pkg/webhook/context"
@@ -32,7 +31,7 @@ import (
 	"github.com/gardener/gardener-extension-shoot-rsyslog-relp/pkg/apis/rsyslog"
 	rsysloginstall "github.com/gardener/gardener-extension-shoot-rsyslog-relp/pkg/apis/rsyslog/install"
 	. "github.com/gardener/gardener-extension-shoot-rsyslog-relp/pkg/webhook/operatingsystemconfig"
-	"github.com/gardener/gardener-extension-shoot-rsyslog-relp/pkg/webhook/operatingsystemconfig/testdata"
+	"github.com/gardener/gardener-extension-shoot-rsyslog-relp/pkg/webhook/operatingsystemconfig/webhooktest"
 )
 
 var _ = Describe("Ensurer", func() {
@@ -151,12 +150,12 @@ var _ = Describe("Ensurer", func() {
 		BeforeEach(func() {
 			ensurer = NewEnsurer(fakeClient, decoder, logger)
 			files = []extensionsv1alpha1.File{oldFile}
-			expectedFiles = append([]extensionsv1alpha1.File{oldFile}, testdata.GetAuditRulesFiles(true)...)
+			expectedFiles = append([]extensionsv1alpha1.File{oldFile}, webhooktest.GetAuditRulesFiles(true)...)
 		})
 
 		Context("when tls is not enabled", func() {
 			BeforeEach(func() {
-				expectedFiles = append(expectedFiles, testdata.GetRsyslogFiles(testdata.GetTestingRsyslogConfig(), true)...)
+				expectedFiles = append(expectedFiles, webhooktest.GetRsyslogFiles(webhooktest.GetTestingRsyslogConfig(), true)...)
 			})
 
 			It("should add additional files to the current ones", func() {
@@ -165,8 +164,8 @@ var _ = Describe("Ensurer", func() {
 			})
 
 			It("should modify already existing rsyslog configuration files", func() {
-				files = append(files, testdata.GetRsyslogFiles(testdata.GetTestingRsyslogConfig(), false)...)
-				files = append(files, testdata.GetAuditRulesFiles(false)...)
+				files = append(files, webhooktest.GetRsyslogFiles(webhooktest.GetTestingRsyslogConfig(), false)...)
+				files = append(files, webhooktest.GetAuditRulesFiles(false)...)
 
 				Expect(ensurer.EnsureAdditionalFiles(ctx, gctx, &files, nil)).To(Succeed())
 				Expect(files).To(ConsistOf(expectedFiles))
@@ -206,8 +205,8 @@ var _ = Describe("Ensurer", func() {
 					PermittedPeer:       []string{"rsyslog-server.foo", "rsyslog-server.foo.bar"},
 				}
 
-				expectedFiles = append(expectedFiles, testdata.GetRsyslogFiles(testdata.GetRsyslogConfigWithTLS(), true)...)
-				expectedFiles = append(expectedFiles, testdata.GetRsyslogTLSFiles(true)...)
+				expectedFiles = append(expectedFiles, webhooktest.GetRsyslogFiles(webhooktest.GetRsyslogConfigWithTLS(), true)...)
+				expectedFiles = append(expectedFiles, webhooktest.GetRsyslogTLSFiles(true)...)
 			})
 
 			It("should add additional files to the current ones", func() {
@@ -216,9 +215,9 @@ var _ = Describe("Ensurer", func() {
 			})
 
 			It("should modify already existing rsyslog configuration files", func() {
-				files = append(files, testdata.GetRsyslogFiles(testdata.GetRsyslogConfigWithTLS(), false)...)
-				files = append(files, testdata.GetAuditRulesFiles(false)...)
-				files = append(files, testdata.GetRsyslogTLSFiles(false)...)
+				files = append(files, webhooktest.GetRsyslogFiles(webhooktest.GetRsyslogConfigWithTLS(), false)...)
+				files = append(files, webhooktest.GetAuditRulesFiles(false)...)
+				files = append(files, webhooktest.GetRsyslogTLSFiles(false)...)
 
 				Expect(ensurer.EnsureAdditionalFiles(ctx, gctx, &files, nil)).To(Succeed())
 				Expect(files).To(ConsistOf(expectedFiles))
@@ -256,7 +255,7 @@ auditRules: |
 					ConfigMapReferenceName: ptr.To("audit-rules"),
 				}
 
-				expectedFiles = append([]extensionsv1alpha1.File{oldFile}, testdata.GetRsyslogFiles(testdata.GetTestingRsyslogConfig(), true)...)
+				expectedFiles = append([]extensionsv1alpha1.File{oldFile}, webhooktest.GetRsyslogFiles(webhooktest.GetTestingRsyslogConfig(), true)...)
 				expectedFiles = append(expectedFiles, []extensionsv1alpha1.File{
 					{
 						Path:        "/var/lib/rsyslog-relp-configurator/audit/rules.d/00_shoot_rsyslog_relp.rules",
@@ -282,7 +281,7 @@ auditRules: |
 				extensionProviderConfig.AuditConfig = &rsyslog.AuditConfig{
 					Enabled: false,
 				}
-				expectedFiles = append([]extensionsv1alpha1.File{oldFile}, testdata.GetRsyslogFiles(testdata.GetTestingRsyslogConfig(), true)...)
+				expectedFiles = append([]extensionsv1alpha1.File{oldFile}, webhooktest.GetRsyslogFiles(webhooktest.GetTestingRsyslogConfig(), true)...)
 			})
 
 			It("should add additional files to the current ones, but not include audit rules files", func() {
@@ -306,7 +305,7 @@ auditRules: |
 		BeforeEach(func() {
 			ensurer = NewEnsurer(fakeClient, decoder, logger)
 			units = []extensionsv1alpha1.Unit{oldUnit}
-			expectedUnits = append([]extensionsv1alpha1.Unit{oldUnit}, testdata.GetRsyslogConfiguratorUnit(true))
+			expectedUnits = append([]extensionsv1alpha1.Unit{oldUnit}, webhooktest.GetRsyslogConfiguratorUnit(true))
 		})
 
 		It("should add additional units to the current ones", func() {
@@ -315,7 +314,7 @@ auditRules: |
 		})
 
 		It("should modify existing units", func() {
-			units = append(units, testdata.GetRsyslogConfiguratorUnit(false))
+			units = append(units, webhooktest.GetRsyslogConfiguratorUnit(false))
 
 			Expect(ensurer.EnsureAdditionalUnits(ctx, nil, &units, nil)).To(Succeed())
 			Expect(units).To(ConsistOf(expectedUnits))
