+++ pkg/webhook/operatingsystemconfig/webhooktest/testdata/00-base-config.rules
