+++ pkg/webhook/operatingsystemconfig/webhooktest/testdata/11-privileged-special.rules
