+++ pkg/webhook/operatingsystemconfig/webhooktest/testdata/12-system-integrity.rules
