+++ pkg/webhook/operatingsystemconfig/webhooktest/testdata/60-audit-with-tls.conf
