+++ pkg/webhook/operatingsystemconfig/webhooktest/testdata/configure-rsyslog.sh
