+++ pkg/webhook/operatingsystemconfig/webhooktest/testdata/process-rsyslog-pstats.sh
