+++ pkg/webhook/operatingsystemconfig/webhooktest/testdata/rsyslog-config-simple.conf.tpl
